refactor: split app settings and diff helpers
